# swagger_docs.json
{"swagger":"2.0","info":{"description":"This is api of LavaJato project","version":"0.1","title":"Lava Jato API","contact":{"name":"LuanGrillo","url":"https://www.luan.dev","email":"contact@luan.dev"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"auth-controller","description":"Auth Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"checklist-controller","description":"Checklist Controller"},{"name":"cidade-controller","description":"Cidade Controller"},{"name":"cliente-controller","description":"Cliente Controller"},{"name":"endereco-controller","description":"Endereco Controller"},{"name":"pedido-controller","description":"Pedido Controller"},{"name":"test-controller","description":"Test Controller"},{"name":"uf-controller","description":"Uf Controller"},{"name":"veiculo-controller","description":"Veiculo Controller"}],"paths":{"/api/auth/signin":{"post":{"tags":["auth-controller"],"summary":"authenticateUser","operationId":"authenticateUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"loginRequest","description":"loginRequest","required":true,"schema":{"$ref":"#/definitions/LoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/auth/signout":{"post":{"tags":["auth-controller"],"summary":"logoutUser","operationId":"logoutUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"summary":"registerUser","operationId":"registerUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"signUpRequest","description":"signUpRequest","required":true,"schema":{"$ref":"#/definitions/SignupRequest"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/test/admin":{"get":{"tags":["test-controller"],"summary":"adminAccess","operationId":"adminAccessUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/test/all":{"get":{"tags":["test-controller"],"summary":"allAccess","operationId":"allAccessUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/test/mod":{"get":{"tags":["test-controller"],"summary":"moderatorAccess","operationId":"moderatorAccessUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/test/user":{"get":{"tags":["test-controller"],"summary":"userAccess","operationId":"userAccessUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/checklist":{"get":{"tags":["checklist-controller"],"summary":"findAll","operationId":"findAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Checklist"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["checklist-controller"],"summary":"insert","operationId":"insertUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Checklist"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Checklist"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/checklist/{id}":{"get":{"tags":["checklist-controller"],"summary":"find","operationId":"findUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Checklist"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["checklist-controller"],"summary":"update","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Checklist"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Checklist"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["checklist-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/cidade":{"get":{"tags":["cidade-controller"],"summary":"findAll","operationId":"findAllUsingGET_1","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Cidade"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["cidade-controller"],"summary":"insert","operationId":"insertUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Cidade"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cidade"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/cidade/findByUf/{idUf}":{"get":{"tags":["cidade-controller"],"summary":"findByUf","operationId":"findByUfUsingGET","produces":["*/*"],"parameters":[{"name":"idUf","in":"path","description":"idUf","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Cidade"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/cidade/{id}":{"get":{"tags":["cidade-controller"],"summary":"find","operationId":"findUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cidade"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["cidade-controller"],"summary":"update","operationId":"updateUsingPUT_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Cidade"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cidade"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["cidade-controller"],"summary":"delete","operationId":"deleteUsingDELETE_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/cliente":{"get":{"tags":["cliente-controller"],"summary":"findAll","operationId":"findAllUsingGET_2","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Cliente"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["cliente-controller"],"summary":"insert","operationId":"insertUsingPOST_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Cliente"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/cliente/{id}":{"get":{"tags":["cliente-controller"],"summary":"find","operationId":"findUsingGET_2","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["cliente-controller"],"summary":"update","operationId":"updateUsingPUT_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Cliente"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["cliente-controller"],"summary":"delete","operationId":"deleteUsingDELETE_2","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/endereco":{"get":{"tags":["endereco-controller"],"summary":"findAll","operationId":"findAllUsingGET_3","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Endereco"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["endereco-controller"],"summary":"insert","operationId":"insertUsingPOST_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Endereco"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Endereco"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/endereco/{id}":{"get":{"tags":["endereco-controller"],"summary":"find","operationId":"findUsingGET_3","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Endereco"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["endereco-controller"],"summary":"update","operationId":"updateUsingPUT_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Endereco"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Endereco"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["endereco-controller"],"summary":"delete","operationId":"deleteUsingDELETE_3","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/pedido":{"get":{"tags":["pedido-controller"],"summary":"findAll","operationId":"findAllUsingGET_4","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Pedido"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["pedido-controller"],"summary":"insert","operationId":"insertUsingPOST_4","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Pedido"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pedido"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/pedido/update_status/{id}":{"put":{"tags":["pedido-controller"],"summary":"updateStatus","operationId":"updateStatusUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Pedido"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pedido"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/pedido/{id}":{"get":{"tags":["pedido-controller"],"summary":"find","operationId":"findUsingGET_4","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pedido"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["pedido-controller"],"summary":"update","operationId":"updateUsingPUT_4","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Pedido"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pedido"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["pedido-controller"],"summary":"delete","operationId":"deleteUsingDELETE_4","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/uf":{"get":{"tags":["uf-controller"],"summary":"findAll","operationId":"findAllUsingGET_5","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Uf"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["uf-controller"],"summary":"insert","operationId":"insertUsingPOST_5","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Uf"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Uf"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/uf/{id}":{"get":{"tags":["uf-controller"],"summary":"find","operationId":"findUsingGET_5","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Uf"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["uf-controller"],"summary":"update","operationId":"updateUsingPUT_5","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Uf"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Uf"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["uf-controller"],"summary":"delete","operationId":"deleteUsingDELETE_5","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/veiculo":{"get":{"tags":["veiculo-controller"],"summary":"findAll","operationId":"findAllUsingGET_6","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Veiculo"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["veiculo-controller"],"summary":"insert","operationId":"insertUsingPOST_6","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Veiculo"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Veiculo"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/veiculo/{id}":{"get":{"tags":["veiculo-controller"],"summary":"find","operationId":"findUsingGET_6","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Veiculo"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["veiculo-controller"],"summary":"update","operationId":"updateUsingPUT_6","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"obj","description":"obj","required":true,"schema":{"$ref":"#/definitions/Veiculo"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Veiculo"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["veiculo-controller"],"summary":"delete","operationId":"deleteUsingDELETE_6","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Checklist":{"type":"object","properties":{"checklistEntrada":{"type":"boolean"},"checklistSaida":{"type":"boolean"},"dataEntrada":{"type":"string","format":"date-time"},"dataSaida":{"type":"string","format":"date-time"},"faroisLanternas":{"type":"boolean"},"id":{"type":"integer","format":"int32"},"interiro_veiculo":{"type":"boolean"},"lataria":{"type":"boolean"},"pedido":{"$ref":"#/definitions/Pedido"},"pneu":{"type":"boolean"},"quilometragem":{"type":"boolean"},"vidro":{"type":"boolean"}},"title":"Checklist"},"Cidade":{"type":"object","required":["id","nome"],"properties":{"uf":{"$ref":"#/definitions/Uf"},"id":{"type":"integer","format":"int32","example":0,"description":"Identificador da cidade"},"nome":{"type":"string","example":"São Paulo","description":"Nome da cidade"}},"title":"Cidade"},"Cliente":{"type":"object","required":["cpf","email","id","nome","telefone"],"properties":{"endereco":{"$ref":"#/definitions/Endereco"},"id":{"type":"integer","format":"int32","example":0,"description":"Identificador do cliente"},"cpf":{"type":"string","example":12345678901,"description":"CPF do cliente"},"nome":{"type":"string","example":"João","description":"Nome do cliente"},"telefone":{"type":"string","example":11999999999,"description":"Telefone do cliente"},"email":{"type":"string","example":"teste@teste.com","description":"Email do cliente"}},"title":"Cliente"},"Endereco":{"type":"object","required":["id"],"properties":{"cliente":{"$ref":"#/definitions/Cliente"},"id":{"type":"integer","format":"int32","example":1,"description":"Identificador do endereço"},"rua":{"type":"string","example":"Rua dos Bobos","description":"Logradouro do endereço"},"cep":{"type":"string","example":12345678,"description":"Número do endereço"},"numero":{"type":"string","example":"Centro","description":"Bairro do endereço"},"bairro":{"type":"string","example":"São Paulo","description":"Cidade do endereço"},"complemento":{"type":"string","example":"SP","description":"Estado do endereço"},"observacoes":{"type":"string","example":"Rua sem saida","description":"OBS"},"cidade":{"example":"SP","description":"Estado do endereço","$ref":"#/definitions/Cidade"}},"title":"Endereco"},"LoginRequest":{"type":"object","properties":{"password":{"type":"string"},"username":{"type":"string"}},"title":"LoginRequest"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Pedido":{"type":"object","required":["cliente","dataInicio","id","observacoes","status","user","valorAcrecimo","valorDesconto"],"properties":{"servico":{"type":"array","items":{"$ref":"#/definitions/Servico"}},"id":{"type":"integer","format":"int32","example":0,"description":"Identificador do pedido"},"valorAcrecimo":{"type":"number","format":"double","example":0.0,"description":"Valor do acréscimo"},"valorDesconto":{"type":"number","format":"double","example":0.0,"description":"Valor do desconto"},"dataInicio":{"type":"string","format":"date-time","example":"2020-01-01","description":"Data de início do pedido"},"observacoes":{"type":"string","example":"Lavagem de carro","description":"Observação do pedido"},"status":{"type":"string","example":"ABERTO","description":"Status do pedido"},"user":{"example":1,"description":"Identificador do usuário","$ref":"#/definitions/User"},"cliente":{"example":1,"description":"Identificador do cliente","$ref":"#/definitions/Cliente"}},"title":"Pedido"},"Roles":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","enum":["ROLE_FUNCIONARIO","ROLE_GERENTE","ROLE_ADMIN"]}},"title":"Roles"},"Servico":{"type":"object","required":["descricao","id","nome","valor"],"properties":{"id":{"type":"integer","format":"int32","example":0,"description":"Identificador do serviço"},"nome":{"type":"string","example":"Lavagem","description":"Nome do serviço"},"valor":{"type":"number","format":"float","example":"R$ 100,00","description":"Valor do serviço"},"descricao":{"type":"string","example":"Lavagem de carro","description":"Descrição do serviço"}},"title":"Servico"},"SignupRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"array","items":{"type":"string"}},"username":{"type":"string"}},"title":"SignupRequest"},"Uf":{"type":"object","required":["id","nome","sigla"],"properties":{"id":{"type":"integer","format":"int32","example":0,"description":"Identificador da UF"},"sigla":{"type":"string","example":"SP","description":"Sigla da UF"},"nome":{"type":"string","example":"São Paulo","description":"Nome da UF"}},"title":"Uf"},"User":{"type":"object","required":["email","id","password","roles","username"],"properties":{"id":{"type":"integer","format":"int64","example":0,"description":"Identificador do usuário"},"username":{"type":"string","example":"Lavajato","description":"Nome do usuário"},"email":{"type":"string","example":"teste@teste.com","description":"Email do usuário"},"password":{"type":"string","example":1234,"description":"Senha do usuário"},"roles":{"type":"array","example":"ROLE_FUNCIONARIO","description":"Permissões do usuário","items":{"$ref":"#/definitions/Roles"}}},"title":"User"},"Veiculo":{"type":"object","required":["ano","cor","id","modelo","placa"],"properties":{"cliente":{"$ref":"#/definitions/Cliente"},"id":{"type":"integer","format":"int32","example":0,"description":"Identificador do veículo"},"placa":{"type":"string","example":"ABC-1234","description":"Placa do veículo"},"modelo":{"type":"string","example":"Fusca","description":"Modelo do veículo"},"cor":{"type":"string","example":"Volkswagen","description":"Marca do veículo"},"ano":{"type":"string","example":1979,"description":"Ano do veículo"}},"title":"Veiculo"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}